requirements.txt
psycopg2==2.6.1
bjoern==1.4.2
falcon==0.3.0
gevent==1.0.2
greenlet==0.4.9
python-mimeparse==0.1.4
six==1.10.0
wheezy.caching==0.1.114
wheezy.core==0.1.140
wheezy.html==0.1.147
wheezy.http==0.1.340
wheezy.routing==0.1.157
wheezy.security==0.1.64
wheezy.validation==0.1.135
wheezy.web==0.1.485
